% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arabidopsis_proteome}
\alias{arabidopsis_proteome}
\title{Arabidopsis thaliana proteome}
\format{A single text file in fasta format.}
\source{
\url{https://www.arabidopsis.org/}
}
\usage{
arabidopsis_proteome
}
\description{
Complete proteome (i.e., all coding genes) of Arabidopsis thaliana.
}
\keyword{datasets}
