% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_alignment_stats.R
\name{calculate_alignment_stats}
\alias{calculate_alignment_stats}
\title{Calculate summary statistics for an alignment.}
\usage{
calculate_alignment_stats(
  alignment,
  cutoff = 120,
  cutoff_any = FALSE,
  include_aln = FALSE
)
}
\arguments{
\item{alignment}{Input alignment; must be a matrix of class "DNAbin".}

\item{cutoff}{Numeric value indicating minimum exon length (optional);
flag this alignment if any/all exons are less than the cutoff length.}

\item{cutoff_any}{Logical; Should the alignment
be flagged if any exons are shorter than the cutoff? The default, FALSE,
means that the alignment will only be flagged if all exons are shorter
than the cutoff value.}

\item{include_aln}{Logical; Should the original alignment
be included in the output list?}
}
\value{
A list including the following summary statistics: \describe{
  \item{intron_lengths}{List including vector of intron lengths}
  \item{exon_lengths}{List including vector of exon lengths}
  \item{num_introns}{Number of introns}
  \item{num_exons}{Number of introns}
  \item{mean_dist}{Mean genetic distance between sequences in alignment}
  \item{max_dist}{Maximum genetic distance between sequences in alignment}
  \item{GC_content}{Total \%GC content}
  \item{pars_inf}{Fraction of sites that are parsimony informative}
  \item{total_exon_length}{Total exon length}
  \item{less_than_cutoff}{Logical flag indicating whether alignment passed
  the minimum exon length cutoff or not}
  \item{alignment}{The original input alignment}
}
}
\description{
Including the original alignment in the output with \code{include_aln}
can be useful for mapping \code{calculate_alignment_stats} over a list
of alignments with \code{\link[purrr]{map_df}} to sort and filter
alignments by their summary statistics.
}
