% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedtools.R
\name{clean_gff}
\alias{clean_gff}
\title{Clean up data from a gff file and
convert to bed format}
\usage{
clean_gff(region, check.chr = FALSE, verbose = FALSE)
}
\arguments{
\item{region}{Dataframe; list of gene regions in "bed" format. Must include
the following columns in order: `chr` ('chromosome', character), `start`
(start position, numeric), and `end` (end position, numeric).}

\item{check.chr}{Logical; should coordinates be checked for chromosomal
format with "chr" prefix?}

\item{verbose}{Logical; should `bedr` functions output all messages?}
}
\value{
Dataframe in "bed" format.
}
\description{
Helper function for `find_bed_regions`. Merges overlapping regions and sorts
regions.
}
