% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yang_and_smith.R
\name{cut_long_internal_branches}
\alias{cut_long_internal_branches}
\title{cut_long_internal_branches}
\usage{

  cut_long_internal_branches(path_to_ys = pkgconfig::get_config("baitfindR::path_to_ys"),
  tree_folder, tree_file_ending, internal_branch_length_cutoff,
  minimal_taxa = 4, outdir, get_hash = FALSE, ...)
}
\arguments{
\item{path_to_ys}{Character vector of length one; the path to the folder containing Y&S python scripts, e.g., \code{"/Users/me/apps/phylogenomic_dataset_construction/"}}

\item{tree_folder}{Character vector of length one; the path to the folder containing the trees to cut.}

\item{tree_file_ending}{Character; string used to match tree files that should be cut.}

\item{internal_branch_length_cutoff}{Numeric vector of length one; the depth at which cuts should be made (smaller numbers indicate greater depth).}

\item{minimal_taxa}{Numeric; minimal number of taxa required for tree to be cut. Default 4, the minimum number of taxa needed for an un-rooted tree.}

\item{outdir}{Character vector of length one; the path to the folder where the subtrees should be written.}

\item{get_hash}{Logical; should the 32-byte MD5 hash be computed for all output subtree files concatenated together? Used for by \code{\link{drake}} for tracking during workflows. If \code{TRUE}, this function will return the hash.}

\item{...}{Other arguments. Not used by this function, but meant to be used by \code{\link{drake}} for tracking during workflows.}
}
\value{
For each input tree with a file ending in \code{tree_file_ending} in \code{tree_folder}, a trimmed tree with a file ending in \code{.mm} will be written to \code{tree_folder}. If \code{get_hash} is \code{TRUE}, the 32-byte MD5 hash be computed for all masked tree files concatenated together will be returned.
}
\description{
Wrapper for Yang and Smith (2014) \code{cut_long_internal_branches.py}
}
\details{
Given a folder containing phylogenetic trees, split the trees into multiple subtrees
for nodes that bifurcate deeper than \code{internal_branch_length_cutoff}.
\code{tree_folder} and \code{outdir} should be different to avoid writing over input trees.
This function will overwrite any output files with the same name in \code{outdir}.
}
\examples{
\dontrun{mask_tips_by_taxonID_transcripts(tree_folder = "some/folder/containing/tree/files", aln_folder = "some/folder/containing/alignment/files")}
}
\references{
Yang, Y. and S.A. Smith. 2014. Orthology inference in non-model organisms using transcriptomes and low-coverage genomes: improving accuracy and matrix occupancy for phylogenomics. Molecular Biology and Evolution 31:3081-3092. \url{https://bitbucket.org/yangya/phylogenomic_dataset_construction/overview}
}
\author{
Joel H Nitta, \email{joelnitta@gmail.com}
}
