% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc.R
\name{delete_old_output}
\alias{delete_old_output}
\title{Delete old output in a folder}
\usage{
delete_old_output(folder, terms)
}
\arguments{
\item{folder}{Path to folder}

\item{terms}{A regular expression: search terms to find files to delete.}
}
\value{
The deleted paths (invisibly).
}
\description{
All file names matching `terms` will be deleted
}
\examples{
\dontrun{
make_dir("test")
fs::file_create("test/a")
fs::file_create("test/b")
fs::file_create("test/1")
list.files("test")
delete_old_output("test", "c")
list.files("test")
delete_old_output("test", "a|b")
list.files("test")
fs::file_delete("test")
}
}
