% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_transcriptomes}
\alias{example_transcriptomes}
\title{Fern transcriptome}
\format{List of lists, each of class 'DNAbin".}
\source{
\url{http://www.onekp.com/public_data.html}
}
\usage{
example_transcriptomes
}
\description{
List of fern transcriptomes from the oneKp project that have been randomly
downsized to 5 percent of original size. Named by 1kP code.
}
\keyword{datasets}
