% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedtools.R
\name{extract_regions_from_fasta}
\alias{extract_regions_from_fasta}
\title{Extract regions from a fasta file}
\usage{
extract_regions_from_fasta(bed_file, fasta_file, out_fasta_file, ...)
}
\arguments{
\item{bed_file}{Path to bed file with locations of regions to extract.
bed file is a tab-separated file with columns for chromosome (e.g., chr1),
start position (e.g., 1), and end position (e.g., 10), in that order.
No column headers are used.}

\item{fasta_file}{Path to file in fasta format to extract regions from.}

\item{out_fasta_file}{Path to write extracted regions (in fasta format).}
}
\value{
List; output of processx::run(). Externally, a fasta file will be
written to the path specified by `out_fasta_file`.
}
\description{
Wrapper for bedtools getfasta.
}
\examples{
extract_regions_from_fasta(
  bed_file = "temp_dir/test_genes"
  fasta_file = "data_raw/Arabidopsis_thaliana.TAIR10.dna.toplevel.renamed.fasta"
  out_fasta_file = "temp_dir/test_masked_genes"
)
}
\author{
Joel H Nitta, \email{joelnitta@gmail.com}
}
