% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_introns.R
\name{fill_introns}
\alias{fill_introns}
\title{Fill-in introns in an alignment}
\usage{
fill_introns(alignment, ref_pattern, outgroup = NULL, trim_outgroup = FALSE)
}
\arguments{
\item{alignment}{matrix of class DNAbin}

\item{ref_pattern}{Pattern used for matching with grep to identify reference
sequences.}

\item{outgroup}{Character vector; names of outgroup sequences.}

\item{trim_outgroup}{Logical; should the outgroups be trimmed from the
alignment?}
}
\value{
Matrix of class DNAbin
}
\description{
The alignment must be produced by aligning sequence to a reference with
introns masked by 'n's, which results in gaps in the rest of the sequences.
This function replaces those gaps with 'n's and removes the reference
sequence.
}
\examples{
library(ape)
data(woodmouse)

# Make reference sequence with a 50bp intron (a string of 'n's)
# in the middle.
woodmouse_ref <- as.character(woodmouse[1,])
woodmouse_ref <- c(woodmouse_ref[1:400],
  rep("n", 50),
  woodmouse_ref[401:length(woodmouse_ref)])
woodmouse_ref <- as.DNAbin(woodmouse_ref)

# Align with other sequences that don't include the intron.
# (need to convert to list first)
woodmouse_ref <- as.list(woodmouse_ref)
names(woodmouse_ref) <- "ref"
woodmouse <- as.list(woodmouse)
woodmouse_with_introns <- ips::mafft(
  c(woodmouse, woodmouse_ref),
  exec = "/usr/bin/mafft")

# Image of the alignment shows that 'ref' has 'n's at positions 400-450,
# while other sequences have gaps ('-').
image(woodmouse_with_introns)

# Fill-in introns
woodmouse_masked <- fill_introns(
  woodmouse_with_introns,
  ref_pattern = "ref"
)

# After filling-in, the reference sequence is gone and
# all introns are 'n's.
image(woodmouse_masked)

}
