% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yang_and_smith.R
\name{filter_1to1_orthologs}
\alias{filter_1to1_orthologs}
\title{filter_1to1_orthologs}
\usage{

  filter_1to1_orthologs(path_to_ys = pkgconfig::get_config("baitfindR::path_to_ys"),
  tree_folder, tree_file_ending, minimal_taxa = 4, outdir,
  overwrite = FALSE, get_hash = TRUE, ...)
}
\arguments{
\item{path_to_ys}{Character vector of length one; the path to the folder containing Y&S python scripts, e.g., "/Users/me/apps/phylogenomic_dataset_construction/"}

\item{tree_folder}{Character vector of length one; the path to the folder containing the trees to be used for filtering.}

\item{tree_file_ending}{Character vector of length one; only tree files with this file ending will be used.}

\item{minimal_taxa}{Numeric; minimal number of taxa required for tree to be included. Default 4, the minimum number of taxa needed for an un-rooted tree.}

\item{outdir}{Character vector of length one; the path to the folder where the filtered trees should be written.}

\item{overwrite}{Logical; should previous output of this command be erased so new output can be written? Once erased it cannot be restored, so use with caution!}

\item{get_hash}{Logical; should the 32-byte MD5 hash be computed for all filtered tree files concatenated together? Used for by \code{\link{drake}} for tracking during workflows. If \code{TRUE}, this function will return the hash.}

\item{...}{Other arguments. Not used by this function, but meant to be used by \code{\link{drake}} for tracking during workflows.}
}
\value{
For each tree file ending in \code{tree_file_ending} in \code{tree_folder}, that tree will be written to \code{outdir} if it consists solely of one-to-one orthologs with the file ending \code{.1to1ortho.tre}. If \code{get_hash} is \code{TRUE}, the 32-byte MD5 hash be computed for all filtered tree files concatenated together will be returned.
}
\description{
Wrapper for Yang and Smith (2014) filter_1to1_orthologs.py
}
\details{
Given a folder containing homolog trees, filter the trees to only those
containing one-to-one orthologs (i.e., no duplications within a sample).
This function will overwrite any output files with the same name in
\code{outdir}.
}
\examples{
\dontrun{filter_1to1_orthologs(tree_folder = "some/folder/containing/tree/files", tree_file_ending = ".tre", tree_folder = "some/folder/containing/tree/files", outdir = "some/folder")}
}
\references{
Yang, Y. and S.A. Smith. 2014. Orthology inference in non-model organisms using transcriptomes and low-coverage genomes: improving accuracy and matrix occupancy for phylogenomics. Molecular Biology and Evolution 31:3081-3092. \url{https://bitbucket.org/yangya/phylogenomic_dataset_construction/overview}
}
\author{
Joel H Nitta, \email{joelnitta@gmail.com}
}
