% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedtools.R
\name{find_bed_regions}
\alias{find_bed_regions}
\title{Find genes, exons, and introns in a gff3 file}
\usage{
find_bed_regions(gff3_file, source_select = NULL, gene_label = "gene",
  exon_label = "exon", verbose = FALSE, prefix = NULL,
  out_dir = NULL, out_type = c("genes", "introns", "exons",
  "write_all"), ...)
}
\arguments{
\item{gff3_file}{Path to input file in `gff3` format.}

\item{source_select}{Character vector; only use regions from these
sources. Must match values in `source` column of gff3 file. Optional.}

\item{gene_label}{String; value used to indicate genes in gff3 file.
Must match at least one value in `type` column of gff3 file. Default "gene".}

\item{exon_label}{String; value used to indicate exons in gff3 file.
Must match at least one value in `type` column of gff3 file. Default "exon".}

\item{verbose}{Logical; should `bedr` functions output all messages?}

\item{prefix}{String; prefix to attach to tsv files if `out_type` is
"write_all".}

\item{out_dir}{Directory to write tsv files if `out_type` is "write_all".}

\item{out_type}{Type of output to return:
"genes": dataframe in "bed" format of genes.
"introns": dataframe in "bed" format of introns.
"exons": dataframe in "bed" format of exons.
"write_all": write tab-separated files for each of `genes`, `introns`, and
`exons` to `out_dir`. The hash digest of the combined genes, introns, and
exons will be returned.}

\item{...}{Other arguments. Not used by this function, but meant to
be used by \code{\link[drake]{drake_plan}} for tracking during workflows.}
}
\value{
Dataframe or character.
}
\description{
If tsv files are written out by selecting "write_all" for `out_type`,
they will overwrite any existing files with the same name in `out_dir`.
}
\examples{
# Find genes

arabidopsis_gff_file <- system.file("extdata", "Arabidopsis_thaliana_TAIR10_40_small.gff3", package = "baitfindR", mustWork = TRUE)

genes <- find_bed_regions(
  gff3_file = arabidopsis_gff_file,
  source_select = "araport11",
  out_type = "genes"
)
tibble::as_tibble(genes)

# Find introns
introns <- find_bed_regions(
  gff3_file = arabidopsis_gff_file,
  source_select = "araport11",
  out_type = "introns"
)
tibble::as_tibble(introns)

# Find exons
exons <- find_bed_regions(
  gff3_file = arabidopsis_gff_file,
  source_select = "araport11",
  out_type = "exons"
)
tibble::as_tibble(exons)

# Write genes, introns, and exons out as tsv files
temp_dir <- tempdir()
find_bed_regions(
  gff3_file = arabidopsis_gff_file,
  source_select = "araport11",
  out_type = "write_all",
  out_dir = temp_dir,
  prefix = "arabidopsis"
)
}
\author{
Joel H Nitta, \email{joelnitta@gmail.com}
}
