% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yang_and_smith.R
\name{fix_names_from_transdecoder}
\alias{fix_names_from_transdecoder}
\title{Shorten names in fasta headers.}
\usage{
fix_names_from_transdecoder(transdecoder_output, mol_type = "dna")
}
\arguments{
\item{transdecoder_output}{Character vector of length one; the path to the .transdecoder.cds file produced by \code{\link{transdecoder_predict}}. It is assumed that the first part of the filename (immediately preceding .transdecoder.cds) is the sample code.}

\item{mol_type}{Character vector of length one; "dna" for DNA or "aa" for proteins.}
}
\value{
Object of class \code{DNAbin} or \code{AAbin} with names shortened to \code{sample_code`"@"`gene}
}
\description{
Does the same thing as Yang and Smith (2014) fix_names_from_transdecoder.py, but works on one fasta file at at time.
}
\examples{
\dontrun{fix_names_from_transdecoder(transdecoder_output = "some/folder/CODE.transdecoder.cds")}
}
\references{
Yang, Y. and S.A. Smith. 2014. Orthology inference in non-model organisms using transcriptomes and low-coverage genomes: improving accuracy and matrix occupancy for phylogenomics. Molecular Biology and Evolution 31:3081-3092. \url{https://bitbucket.org/yangya/phylogenomic_dataset_construction/overview}
}
\author{
Joel H Nitta, \email{joelnitta@gmail.com}
}
