% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc.R
\name{get_out_hash}
\alias{get_out_hash}
\title{Get MD5 hash of all files in a folder}
\usage{
get_out_hash(folder, terms)
}
\arguments{
\item{folder}{Path to folder}

\item{terms}{A regular expression: search terms to find files to digest.}
}
\value{
Character vector of length 1; the MD5 hash.
}
\description{
All file names matching `terms` will included
}
\examples{
\dontrun{
make_dir("test")
fs::file_create("test/a")
fs::file_create("test/b")
fs::file_create("test/1")
list.files("test")
get_out_hash("test", "c")
get_out_hash("test", "a|b")
get_out_hash("test", "d")
get_out_hash("test", "e")
fs::file_delete("test")
}
}
