% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lygodium_proteome}
\alias{lygodium_proteome}
\title{Lygodium japonicum proteome}
\format{A single text file in fasta format.}
\source{
\url{http://bioinf.mind.meiji.ac.jp/kanikusa/}
}
\usage{
lygodium_proteome
}
\description{
Complete proteome (i.e., all coding genes) of Lygodium japonicum. Duplicate sequence IDs were identified and removed.
}
\keyword{datasets}
