% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yang_and_smith.R
\name{mafft_wrapper}
\alias{mafft_wrapper}
\title{mafft_wrapper}
\usage{

  mafft_wrapper(path_to_ys = pkgconfig::get_config("baitfindR::path_to_ys"),
  fasta_folder, infile_ending = "fa", number_cores, seq_type = "dna",
  overwrite = FALSE, get_hash = TRUE, ...)
}
\arguments{
\item{path_to_ys}{Character vector of length one; the path to the folder containing Y&S python scripts, e.g., "/Users/me/apps/phylogenomic_dataset_construction/"}

\item{fasta_folder}{Character vector of length one; the path to the folder containing the fasta files to be aligned.}

\item{infile_ending}{Character vector of length one; only files with this ending will be included.}

\item{number_cores}{Numeric; number of threads to use for and \code{mafft}.}

\item{seq_type}{Character vector of length one indicating type of sequences. Should either be \code{"dna"} for DNA or \code{"aa"} for proteins.}

\item{overwrite}{Logical; should previous output of this command be erased so new output can be written? Once erased it cannot be restored, so use with caution!}

\item{get_hash}{Logical; should the 32-byte MD5 hash be computed for all aligned fasta files concatenated together? Used for by \code{\link[drake]{drake_plan}} for tracking during workflows. If \code{TRUE}, this function will return the hash.}

\item{...}{Other arguments. Not used by this function, but meant to be used by \code{\link[drake]{drake_plan}} for tracking during workflows.}
}
\value{
Aligned fasta files will be written to \code{fasta_folder} with the file ending \code{.aln}. If \code{get_hash} is \code{TRUE}, the 32-byte MD5 hash be computed for all \code{.aln} files concatenated together will be returned.
}
\description{
Wrapper for Yang and Smith (2014) mafft_wrapper.py
}
\details{
Given a directory containing unaligned fasta files, align all fasta files in the directory. If there are > 1000 sequences in the directory, use the mafft \code{--auto} algorithm. If less, use the \code{--genafpair} algorithm.
}
\examples{
\dontrun{mafft_wrapper(fasta_folder = "some/folder/with/fasta/files", number_cores = 2, seq_type = "dna")}
}
\references{
Yang, Y. and S.A. Smith. 2014. Orthology inference in non-model organisms using transcriptomes and low-coverage genomes: improving accuracy and matrix occupancy for phylogenomics. Molecular Biology and Evolution 31:3081-3092. \url{https://bitbucket.org/yangya/phylogenomic_dataset_construction/overview}
}
\author{
Joel H Nitta, \email{joelnitta@gmail.com}
}
