% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yang_and_smith.R
\name{mask_tips_by_taxonID_transcripts}
\alias{mask_tips_by_taxonID_transcripts}
\title{Mask tips in tree.}
\usage{

  mask_tips_by_taxonID_transcripts(path_to_ys = pkgconfig::get_config("baitfindR::path_to_ys"),
  tree_folder, aln_folder, mask_paraphyletic = TRUE, overwrite = FALSE,
  get_hash = TRUE, echo = pkgconfig::get_config("baitfindR::echo",
  fallback = FALSE), ...)
}
\arguments{
\item{path_to_ys}{Character vector of length one; the path to the folder containing Y&S python scripts, e.g., \code{"/Users/me/apps/phylogenomic_dataset_construction/"}}

\item{tree_folder}{Character vector of length one; the path to the folder containing the trees to mask.}

\item{aln_folder}{Character vector of length one; the path to the folder containing the alignments used to make the trees.}

\item{mask_paraphyletic}{Logical; should paraphyletic tips belonging to the same taxon be masked?}

\item{overwrite}{Logical; should previous output of this command be erased so new output can be written? Once erased it cannot be restored, so use with caution!}

\item{get_hash}{Logical; should the 32-byte MD5 hash be computed for all output masked tree files concatenated together? Used for by \code{\link[drake]{drake_plan}} for tracking during workflows. If \code{TRUE}, this function will return the hash.}

\item{echo}{Logical; should the standard output and error be printed to the screen?}

\item{...}{Other arguments. Not used by this function, but meant to be used by \code{\link[drake]{drake_plan}} for tracking during workflows.}
}
\value{
For each input tree with a file ending in \code{.tt} in \code{tree_folder}, a trimmed tree with a file ending in \code{.mm} will be written to \code{tree_folder}. If \code{get_hash} is \code{TRUE}, the 32-byte MD5 hash be computed for all masked tree files concatenated together will be returned.
}
\description{
Given a folder containing phylogenetic trees and their alignments, mask monophyletic
(and optionally, paraphyletic) tips belonging to the same taxon (i.e., keep only a single tip
to represent clades consisting of a single taxon). Tree files are assumed to end in \code{.tt}
(the output of \code{trim_tips}), and only tree files with this ending will be included.
Alignment files are assumed to end in \code{.aln-cln} (the output of \code{fasta_to_tree}),
and only alignment files with this ending will be included.
The tip with the fewest ambiguous characters in the alignment will be kept. This
function will overwrite any output files with the same name in \code{tree_folder}.
}
\details{
Wrapper for Yang and Smith (2014) \code{mask_tips_by_taxonID_transcripts.py}
}
\examples{
\dontrun{mask_tips_by_taxonID_transcripts(tree_folder = "some/folder/containing/tree/files", aln_folder = "some/folder/containing/alignment/files")}
}
\references{
Yang, Y. and S.A. Smith. 2014. Orthology inference in non-model organisms using transcriptomes and low-coverage genomes: improving accuracy and matrix occupancy for phylogenomics. Molecular Biology and Evolution 31:3081-3092. \url{https://bitbucket.org/yangya/phylogenomic_dataset_construction/overview}
}
\author{
Joel H Nitta, \email{joelnitta@gmail.com}
}
