% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcl.R
\name{mcl}
\alias{mcl}
\title{mcl}
\usage{
mcl(mcl_input, mcl_output, i_value, e_value, other_args = NULL, ...)
}
\arguments{
\item{mcl_input}{Character vector of length one; the path to the input file for mcl clustering.}

\item{mcl_output}{Character vector of length one; the path to the output file produced by the mcl algorithm.}

\item{i_value}{Numeric or character vector of length one; the inflation value.}

\item{e_value}{Numeric or character vector of length one; the minimal -log transformed evalue to be considered by the algorithm.}

\item{other_args}{Character vector; other arguments to pass to mcl. Each should be an element of the vector. For example, to pass "-abc" to specify the input file format and "--te" to specify number of threads, use \code{c("--abc", "-te", "2")}.}

\item{...}{Other arguments. Not used by this function, but meant to be used by \code{\link{drake}} for tracking during workflows.}
}
\value{
A plain text file of tab-separated values, where each value on a line belongs to the same cluster.
}
\description{
Wrapper to call mcl
}
\details{
Calls the Markov Cluster Algorithm (mcl), a clustering algorithm for graphs. Here, it is meant to be used on genetic distances from BLAST.
}
\examples{
\dontrun{mcl(mcl_input = "some/folder/distance.file", mcl_output = "some/folder/mcl_output.txt", i_value = 1.4, evalue = 5)}
}
\references{
Stijn van Dongen, A cluster algorithm for graphs. Technical Report INS-R0010, National Research Institute for Mathematics and Computer Science in the Netherlands, Amsterdam, May 2000. \url{https://micans.org/mcl/}
}
\author{
Joel H Nitta, \email{joelnitta@gmail.com}
}
