% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{onekp_data}
\alias{onekp_data}
\title{Taxonomic data for fern transcriptomes}
\format{
A data frame with 8 rows and 6 variables:
\describe{
  \item{code}{4 letter code used to identify the sample}
  \item{species}{Binomial species name}
  \item{genus}{Genus}
  \item{specific_epithet}{Specific epithet}
  \item{family}{Family}
  \item{group}{Ingroup or outgroup status}
}
}
\source{
\url{http://www.onekp.com/public_data.html}

Pteridophyte Phylogeny Group I. 2016. A community-derived classification for extant lycophytes and ferns. Journal of Systematics and Evolution 54:563–603.
}
\usage{
onekp_data
}
\description{
Taxonomic data for seven fern transcripomes downloaded from the onekp project. Family-level taxonomy follows PPG I (2016).
}
\keyword{datasets}
