% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{onekp_ferns}
\alias{onekp_ferns}
\title{Fern transcriptomes}
\format{A list of eight fern transcriptomes stored in binary DNA format (class ape::DNAbin).
All transcriptomes have been randomly subsampled to 2\% of their original size.}
\source{
\url{http://www.onekp.com/public_data.html}
}
\usage{
onekp_ferns
}
\description{
Selected, downsized fern transcriptomes from the 1KP dataset to use for testing code.
}
\keyword{datasets}
