% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PSKY}
\alias{PSKY}
\title{Fern transcriptome}
\format{List of class 'DNAbin".}
\source{
\url{http://www.onekp.com/public_data.html}
}
\usage{
PSKY
}
\description{
A fern transcriptome from the oneKp project that has been randomly downsized
to 5 percent of its original size. Named by 1kP code.
}
\keyword{datasets}
