% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc.R
\name{set_ys_path}
\alias{set_ys_path}
\title{set_ys_path}
\usage{
set_ys_path(path)
}
\arguments{
\item{path}{Character vector of length one; the complete path to the folder containing Y&S python scripts, e.g., \code{"/Users/me/apps/phylogenomic_dataset_construction/"}}
}
\description{
Set the default path to the folder containing Yang and Smith (2014) python scripts.
}
\details{
This will overwrite the default value of the \code{path_to_ys} argument for all \code{baitfindR} functions.
}
\examples{
set_ys_path("/Users/me/apps/phylogenomic_dataset_construction/")
}
\references{
Yang, Y. and S.A. Smith. 2014. Orthology inference in non-model organisms using transcriptomes and low-coverage genomes: improving accuracy and matrix occupancy for phylogenomics. Molecular Biology and Evolution 31:3081-3092. \url{https://bitbucket.org/yangya/phylogenomic_dataset_construction/overview}
}
\author{
Joel H Nitta, \email{joelnitta@gmail.com}
}
