% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_coffee.R
\name{tcs}
\alias{tcs}
\title{tcs}
\usage{
tcs(alignment, number_cores = 1, other_args = NULL, wd = here::here(),
  ...)
}
\arguments{
\item{alignment}{Character vector of length one; the path to the input alignment}

\item{number_cores}{Number of cores to use by t_coffee}

\item{other_args}{Character vector; other arguments to pass to t_coffee. Each should be an element of the vector.}

\item{wd}{Character vector of length one; the directory where the command will be run, and the external output written.}

\item{...}{Additional other arguments. Not used by this function, but meant to be used by \code{\link[drake]{drake_plan}} for tracking during workflows.}
}
\value{
A numeric value; the overall TCS score of the input alignment.

Externally, a plain text file with TCS scores for each sequence, position, and the whole alignment. The file will be named \code{<alignment>.score_ascii}, where \code{<alignment>} is the filename of the input alignment.
}
\description{
Wrapper to call t_coffee to calculate Transitive Consistency Score (TCS)
}
\examples{
\dontrun{tcs("some/alignment.fasta")}

}
\references{
Chang JM, Di Tommaso P, Notredame C. TCS: a new multiple sequence alignment reliability measure to estimate alignment accuracy and improve phylogenetic tree reconstruction. Mol Biol Evol. 2014 Jun. 31(6):1625-37. \url{http://www.tcoffee.org/Projects/tcs/}
}
\author{
Joel H Nitta, \email{joelnitta@gmail.com}
}
