% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_coffee.R
\name{tcs_loop}
\alias{tcs_loop}
\title{tcs_loop}
\usage{
tcs_loop(folder, pattern = NULL, number_cores = 1, other_args = NULL, ...)
}
\arguments{
\item{folder}{Character vector of length one; the path to the folder containing the alignments.}

\item{pattern}{An optional regular expression. Only alignment files with names that match the regular expression will be included.}

\item{number_cores}{Number of cores for t-coffee to use.}

\item{other_args}{Character vector; other arguments to pass to t_coffee. Each should be an element of the vector.}

\item{...}{Additional other arguments. Not used by this function, but meant to be used by \code{\link[drake]{drake_plan}} for tracking during workflows.}
}
\value{
A numeric vector; the overall TCS scores for all alignments in the folder.

Externally, plain text files that include TCS scores for each sequence, position, and the whole alignment overall, for each alignment. The files will be written to \code{folder}, and named \code{<alignment>.score_ascii}, where \code{<alignment>} is the filename of the input alignment.
}
\description{
Call t_coffee to calculate Transitive Consistency Scores (TCS) for all alignments in a folder
}
\examples{
\dontrun{tcs_loop(folder = "some/folder/with/alignments/", pattern = "fa.mafft.aln-cln$")}

}
\references{
Chang JM, Di Tommaso P, Notredame C. TCS: a new multiple sequence alignment reliability measure to estimate alignment accuracy and improve phylogenetic tree reconstruction. Mol Biol Evol. 2014 Jun. 31(6):1625-37. \url{http://www.tcoffee.org/Projects/tcs/}
}
\author{
Joel H Nitta, \email{joelnitta@gmail.com}
}
