% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transdecoder.R
\name{transdecoder_long_orfs}
\alias{transdecoder_long_orfs}
\title{transdecoder_long_orfs}
\usage{

  transdecoder_long_orfs(path_to_transdecoder = pkgconfig::get_config("baitfindR::path_to_transdecoder"),
  transcriptome, wd, other_args = NULL, ...)
}
\arguments{
\item{path_to_transdecoder}{Character vector of length one; the path to the folder containing TransDecoder scripts, e.g., \code{"/Users/me/apps/TransDecoder/"}}

\item{transcriptome}{Character vector of length one; the path to the fasta file containing transcript sequences (i.e., the transcriptome).}

\item{wd}{Character vector of length one; the directory where the command will be run, and the output folder created.}

\item{other_args}{Character vector; other arguments to pass to TransDecoder. Each should be an element of the vector.}

\item{...}{Additional other arguments. Not used by this function, but meant to be used by \code{\link{drake}} for tracking during workflows.}
}
\value{
An output folder in the working directory containing base frequencies and .cds, .gff3, and .pep files for the recovered ORFs. The folder will be named \code{<transcriptome>.transdecoder_dir}, where \code{<transcriptome>} is the value of that argument.
}
\description{
Wrapper to call TransDecoder.LongOrfs
}
\details{
Extracts long open reading frames (ORFs) from a fasta file containing transcript sequences (i.e., the transcriptome).
}
\examples{
\dontrun{transdecoder_long_orfs("some/transcriptome_file.fa")}
}
\references{
\url{http://transdecoder.github.io}
}
\author{
Joel H Nitta, \email{joelnitta@gmail.com}
}
