% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_introns.R
\name{trim_both_ends}
\alias{trim_both_ends}
\title{Trim both ends of alignment}
\usage{
trim_both_ends(alignment)
}
\arguments{
\item{alignment}{matrix of class DNAbin}
}
\value{
matrix of class DNAbin
}
\description{
Trim both ends of alignment
}
\examples{
library(ape)
dna_mat <- matrix(rep(c("n", "t", "a", "n"), 2), 2, 4, byrow = TRUE)
dna_mat <- as.DNAbin(dna_mat)
dna_mat
trim_both_ends(dna_mat)
}
