% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_introns.R
\name{trim_one_end}
\alias{trim_one_end}
\title{Trim columns of exclusively "n" characters from one end of alignment}
\usage{
trim_one_end(alignment)
}
\arguments{
\item{alignment}{matrix of class DNAbin}
}
\value{
matrix of class DNAbin
}
\description{
Trim columns of exclusively "n" characters from one end of alignment
}
\examples{
library(ape)
dna_mat <- matrix(rep(c("n", "t", "a", "n"), 2), 2, 4, byrow = TRUE)
dna_mat
dna_mat <- as.DNAbin(dna_mat)
dna_mat
trim_one_end(dna_mat)
}
