% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc.R
\name{write_fasta_files}
\alias{write_fasta_files}
\title{Write out a list of fasta files to a directory}
\usage{
write_fasta_files(fasta_list, out_dir, fasta_names = NULL, ext = NULL,
  get_hash = TRUE, ...)
}
\arguments{
\item{fasta_list}{List of DNA sequences to write out. Each item in list
must of class \code{\link[ape]{DNAbin}}.}

\item{out_dir}{Path to directory to write out DNA sequences.}

\item{fasta_names}{Optional character vector of file names to use when
writing out DNA sequences.}

\item{ext}{Optional character vector of length one; file extension to
append to DNA sequences (e.g., "fasta").}

\item{get_hash}{Logical; should the MD5 hash of `fasta_list` be returned?}

\item{...}{Additional other arguments. Not used by this function,
but meant to be used by \code{\link[drake]{drake_plan}} for tracking
during workflows.}
}
\value{
None (invisible ‘NULL’) or character vector if `hash` is `TRUE`.
Externally, fasta files will be written to `out_dir`.
}
\description{
Optionally assign names to the files if `fasta_list` isn't already named
or if you want to over-write the original names.
}
\examples{
# Load some example DNA sequences.
library(ape)
data(woodmouse)

# Make a temporary working directory to write out files.
temp_dir <- fs::dir_create(fs::path(tempdir(), "baitfindR_example"))

# Make list of DNA samples.
dna_list <- list(a = woodmouse, b = woodmouse)

# Write out list:
# names as-is
write_fasta_files(dna_list, temp_dir)
list.files(temp_dir)

# add extension
write_fasta_files(dna_list, temp_dir, ext = "fasta")
list.files(temp_dir)

# new names and extension
write_fasta_files(
  dna_list,
  temp_dir,
  fasta_names = c("ho", "ge"),
  ext = "fasta")
list.files(temp_dir)

# Cleanup
fs::file_delete(temp_dir)
}
