% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yang_and_smith.R
\name{write_fasta_files_from_mcl}
\alias{write_fasta_files_from_mcl}
\title{write_fasta_files_from_mcl}
\usage{

  write_fasta_files_from_mcl(path_to_ys = pkgconfig::get_config("baitfindR::path_to_ys"),
  all_fasta, mcl_outfile, minimal_taxa = 4, outdir)
}
\arguments{
\item{path_to_ys}{Character vector of length one; the path to the folder containing Y&S python scripts, e.g., "/Users/me/apps/phylogenomic_dataset_construction/"}

\item{all_fasta}{Character vector of length one; the path to the fasta file including all query sequences concatenated together, i.e., the fasta file used to create the "all-by-all" blast database.}

\item{mcl_outfile}{Character vector of length one; the path to the output from running mcl on blast distances.}

\item{minimal_taxa}{Numeric; minimal number of taxa required to be present or the clus. Default 4, the minimum number of taxa needed for an un-rooted tree.}

\item{outdir}{Character vector of length one; the path to the folder where the clusters should be written.}
}
\description{
Wrapper for Yang and Smith (2014) write_fasta_files_from_mcl.py
}
\details{
Given the output from the mcl clustering algorthim and a concatenated fasta file
including all sequences used for clustering, outputs one fasta file per cluster
including the sequences in that cluster.
}
\examples{
\dontrun{write_fasta_files_from_mcl(all_fasta = "some/folder/all.fasta", mcl_outfile = "some/folder/hit-frac0.4_I1.4_e5", minimal_taxa = 5, outdir = "some/folder")}
}
\references{
Yang, Y. and S.A. Smith. 2014. Orthology inference in non-model organisms using transcriptomes and low-coverage genomes: improving accuracy and matrix occupancy for phylogenomics. Molecular Biology and Evolution 31:3081-3092. \url{https://bitbucket.org/yangya/phylogenomic_dataset_construction/overview}
}
\author{
Joel H Nitta, \email{joelnitta@gmail.com}
}
