% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yang_and_smith.R
\name{write_fasta_files_from_trees}
\alias{write_fasta_files_from_trees}
\title{Write fasta files from trees.}
\usage{
write_fasta_files_from_trees(
  path_to_ys = pkgconfig::get_config("baitfindR::path_to_ys"),
  all_fasta,
  tree_folder,
  tree_file_ending,
  outdir,
  overwrite = FALSE,
  get_hash = TRUE,
  echo = pkgconfig::get_config("baitfindR::echo", fallback = FALSE),
  ...
)
}
\arguments{
\item{path_to_ys}{Character vector of length one; the path to the folder containing Y&S python scripts, e.g., "/Users/me/apps/phylogenomic_dataset_construction/"}

\item{all_fasta}{Character vector of length one; the path to the fasta file including all the sequences that were originally used to build the trees.}

\item{tree_folder}{Character vector of length one; the path to the folder containing the trees to be used for extracting fasta sequences.}

\item{tree_file_ending}{Character vector of length one; only tree files with this file ending will be used.}

\item{outdir}{Character vector of length one; the path to the folder where the fasta files should be written.}

\item{overwrite}{Logical; should previous output of this command be erased so new output can be written? Once erased it cannot be restored, so use with caution!}

\item{get_hash}{Logical; should the 32-byte MD5 hash be computed for all output fasta files concatenated together? Used for by \code{\link[drake]{drake_plan}} for tracking during workflows. If \code{TRUE}, this function will return the hash.}

\item{echo}{Logical; should the standard output and error be printed to the screen?}

\item{...}{Other arguments. Not used by this function, but meant to be used by \code{\link[drake]{drake_plan}} for tracking during workflows.}
}
\value{
One fasta file per tree file ending in \code{tree_file_ending} in \code{tree_folder} will be written to \code{outdir}. If \code{get_hash} is \code{TRUE}, the 32-byte MD5 hash be computed for all output fasta files concatenated together will be returned.
}
\description{
Given a folder containing phylogenetic trees and a single concatenated fasta file
including all the sequences used to build the trees, output one fasta file per tree
with the sequences in that tree. This function will overwrite any output files with
the same name in \code{outdir}.
}
\details{
Wrapper for Yang and Smith (2014) write_fasta_files_from_trees.py
}
\examples{
\dontrun{write_fasta_files_from_trees(all_fasta = "some/folder/all.fasta", tree_file_ending = ".subtree", tree_folder = "some/folder/containing/tree/files", outdir = "some/folder")}
}
\references{
Yang, Y. and S.A. Smith. 2014. Orthology inference in non-model organisms using transcriptomes and low-coverage genomes: improving accuracy and matrix occupancy for phylogenomics. Molecular Biology and Evolution 31:3081-3092. \url{https://bitbucket.org/yangya/phylogenomic_dataset_construction/overview}
}
\author{
Joel H Nitta, \email{joelnitta@gmail.com}
}
