% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yang_and_smith.R
\name{write_ortholog_fasta_files}
\alias{write_ortholog_fasta_files}
\title{write_ortholog_fasta_files}
\usage{

  write_ortholog_fasta_files(path_to_ys = pkgconfig::get_config("baitfindR::path_to_ys"),
  all_fasta, tree_folder, outdir, minimal_taxa = 4, overwrite = FALSE,
  get_hash = TRUE, ...)
}
\arguments{
\item{path_to_ys}{Character vector of length one; the path to the folder containing Y&S python scripts, e.g., "/Users/me/apps/phylogenomic_dataset_construction/"}

\item{all_fasta}{Character vector of length one; the path to the fasta file including all the sequences that were originally used to build the trees.}

\item{tree_folder}{Character vector of length one; the path to the folder containing the trees to be used for extracting sequences.}

\item{outdir}{Character vector of length one; the path to the folder where the fasta files should be written.}

\item{minimal_taxa}{Numeric; minimal number of taxa required for output sequences to be written (regardless of ingroup/outgroup status).}

\item{overwrite}{Logical; should previous output of this command be erased so new output can be written? Once erased it cannot be restored, so use with caution!}

\item{get_hash}{Logical; should the 32-byte MD5 hash be computed for all fasta files concatenated together? Used for by \code{\link{drake}} for tracking during workflows. If \code{TRUE}, this function will return the hash.}

\item{...}{Other arguments. Not used by this function, but meant to be used by \code{\link{drake}} for tracking during workflows.}
}
\value{
One fasta file per tree will be written to \code{outdir}. If \code{get_hash} is \code{TRUE}, the 32-byte MD5 hash be computed for all \code{.fa} files concatenated together will be returned.
}
\description{
Wrapper for Yang and Smith (2014) write_ortholog_fasta_files.py
}
\details{
Given a folder containing ortholog trees, write out the fasta files
that correspond to the sequences in the trees.
}
\examples{
\dontrun{write_ortholog_fasta_files(all_fasta = "some/folder/all.fasta", tree_folder = "some/folder/containing/tree/files", outdir = "some/folder", minimal_taxa = 5)}
}
\references{
Yang, Y. and S.A. Smith. 2014. Orthology inference in non-model organisms using transcriptomes and low-coverage genomes: improving accuracy and matrix occupancy for phylogenomics. Molecular Biology and Evolution 31:3081-3092. \url{https://bitbucket.org/yangya/phylogenomic_dataset_construction/overview}
}
\author{
Joel H Nitta, \email{joelnitta@gmail.com}
}
