% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{blast_n}
\alias{blast_n}
\title{Run a blastn query.}
\usage{
blast_n(query, database, out_file,
  outfmt = "'6 qseqid qlen sseqid slen frames pident nident length mismatch gapopen qstart qend sstart send evalue bitscore'",
  other_args = NULL, ...)
}
\arguments{
\item{query}{Character vector of length one; the path to the fasta
file to use as the query sequence(s).}

\item{database}{Character vector of length one; the name of the blast
database.}

\item{out_file}{Character vector of length one; the name to use for
the results file.}

\item{outfmt}{Character vector of length one; value to pass to
\code{blastn} \code{outfmt} argument. Default = "6".}

\item{other_args}{Character vector; other arguments to pass on to
\code{blastn}. For a list of options, run \code{blastn -help}.}

\item{...}{Additional other arguments. Not used by this function,
but meant to be used by \code{\link[drake]{drake_plan}} for tracking
during workflows.}
}
\value{
A tab-separated text file with the results of the blastn
search, named with the value of \code{out_file}.
}
\description{
This is a wrapper for blastn.
}
\references{
\url{https://www.ncbi.nlm.nih.gov/books/NBK279690/}
}
\author{
Joel H Nitta, \email{joelnitta@gmail.com}
}
