% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{blast_p}
\alias{blast_p}
\title{blast_p}
\usage{
blast_p(query, database, out_file, other_args = NULL, ...)
}
\arguments{
\item{query}{Character vector of length one; the path to the fasta file to use as the query sequence(s).}

\item{database}{Character vector of length one; the name of the blast database.}

\item{out_file}{Character vector of length one; the name to use for the results file.}

\item{other_args}{Character vector; other arguments to pass on to \code{makeblastdb}. For a list of options, run \code{blastp -help}.}

\item{...}{Additional other arguments. Not used by this function, but meant to be used by \code{\link[drake]{drake_plan}} for tracking during workflows.}
}
\value{
A tab-separated text file with the results of the blastp search, named with the value of \code{out_file}.
}
\description{
Wrapper to call blastp
}
\references{
\url{https://www.ncbi.nlm.nih.gov/books/NBK279690/}
}
\author{
Joel H Nitta, \email{joelnitta@gmail.com}
}
