% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{build_blast_db}
\alias{build_blast_db}
\title{Build a BLAST database.}
\usage{
build_blast_db(in_seqs, db_type = "nucl", out_name = NULL,
  title = NULL, parse_seqids = FALSE, echo = TRUE, wd, ...)
}
\arguments{
\item{in_seqs}{Character vector of length one; the path to the fasta
file containing the sequences to be included in the database.}

\item{db_type}{Character vector of length one; "nucl" for DNA or "prot"
for amino acids (proteins).}

\item{out_name}{Character vector of length one; name of BLAST database
to be created (optional). This will be used to name all database files;
if omitted, the name of the `in_seqs` file will be used instead.}

\item{title}{Character vector of length one; title of BLAST database
to be created (optional).}

\item{parse_seqids}{Logical; should the makeblastdb flag
"parse_seqids" be used?}

\item{echo}{Logical; should standard error and output be printed?}

\item{wd}{Character vector of length one; working directory. The blast
database will be made here.}

\item{...}{Additional other arguments. Not used by this function, but
meant to be used by \code{\link[drake]{drake_plan}} for tracking
during workflows.}
}
\value{
A series of files starting with \code{out_name} and ending in
.phr, .pin, .pog, .psd, .psi, psq (for proteins) or .nhr, .nin, .nog,
.nsd, .nsi, and .nsq (for DNA) that constitute the BLAST database in
the working directory.
}
\description{
This is a wrapper for makeblastdb.
}
\examples{
library(ape)
data(woodmouse)
temp_dir <- fs::dir_create(fs::path(tempdir(), "baitfindR_example"))
ape::write.FASTA(woodmouse, fs::path(temp_dir, "woodmouse.fasta"))
list.files(temp_dir)
build_blast_db(
  fs::path(temp_dir, "woodmouse.fasta"),
  title = "test db",
  out_name = "wood",
  parse_seqids = TRUE,
  wd = temp_dir)
list.files(temp_dir)
fs::file_delete(temp_dir)
}
\references{
\url{https://www.ncbi.nlm.nih.gov/books/NBK279690/}
}
\author{
Joel H Nitta, \email{joelnitta@gmail.com}
}
