% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yang_and_smith.R
\name{fasta_to_tree}
\alias{fasta_to_tree}
\title{Infer trees from fasta files.}
\usage{

  fasta_to_tree(path_to_ys = pkgconfig::get_config("baitfindR::path_to_ys"),
  seq_folder, number_cores, seq_type, bootstrap = FALSE,
  overwrite = FALSE, get_hash = TRUE, echo = FALSE, ...)
}
\arguments{
\item{path_to_ys}{Character vector of length one; the path to the folder containing Y&S python scripts, e.g., \code{"/Users/me/apps/phylogenomic_dataset_construction/"}}

\item{seq_folder}{Character vector of length one; the path to the folder containing the fasta files.}

\item{number_cores}{Numeric; number of threads to use for \code{RAxML} and \code{mafft}.}

\item{seq_type}{Character vector of length one indicating type of sequences. Should either be \code{"dna"} for DNA or \code{"aa"} for proteins.}

\item{bootstrap}{Logical; should run a bootstrap analysis be run for the trees?}

\item{overwrite}{Logical; should previous output of this command be erased so new output can be written? Once erased it cannot be restored, so use with caution!}

\item{get_hash}{Logical; should the 32-byte MD5 hash be computed for all output tree files concatenated together? Used for by \code{\link[drake]{drake_plan}} for tracking during workflows. If \code{TRUE}, this function will return the hash.}

\item{echo}{Logical; should the standard output and error be printed to the screen?}

\item{...}{Other arguments. Not used by this function, but meant to be used by \code{\link[drake]{drake_plan}} for tracking during workflows.}
}
\value{
For each input cluster \code{cluster1.fa} in \code{seq_folder}, \code{cluster1.fa.mafft.aln} (small clusters) or \code{cluster1.pasta.aln} (large clusters), \code{cluster1.fa.mafft.aln-cln} (small clusters) or \code{cluster1.fa.pasta.aln-cln} (large clusters), and \code{cluster1.raxml.tre} (small clusters) or \code{cluster1.fasttree.tre} (large clusters) will be written to \code{seq_folder}. If \code{get_hash} is \code{TRUE}, the 32-byte MD5 hash be computed for all \code{.tre} files concatenated together will be returned.
}
\description{
Given a folder containing unaligned sequences in fasta format (i.e., clusters),
aligns each cluster with \code{mafft} (small clusters) or \code{pasta} (large clusters),
excludes poorly aligned sites with \code{phyutility}, and infers a maximum-likelihood
tree with \code{RAxML} (small clusters) or \code{fasttree} (large clusters). Requires all
of these programs to be installed and included in the user's \code{$PATH}. Assumes clusters are named \code{"cluster1.fa"}, \code{"cluster2.fa"}, etc. Clusters with fewer than 1,000 sequences are considered "small," and those with more are considered "large."
}
\details{
Wrapper for Yang and Smith (2014) \code{fasta_to_tree.py}
}
\examples{
\dontrun{fasta_to_tree(seq_folder = "some/folder/containing/fasta/seqs", number_cores = 1, seq_type = "dna", bootstrap = FALSE)}
}
\references{
Yang, Y. and S.A. Smith. 2014. Orthology inference in non-model organisms using transcriptomes and low-coverage genomes: improving accuracy and matrix occupancy for phylogenomics. Molecular Biology and Evolution 31:3081-3092. \url{https://bitbucket.org/yangya/phylogenomic_dataset_construction/overview}
}
\author{
Joel H Nitta, \email{joelnitta@gmail.com}
}
