% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yang_and_smith.R
\name{trim_tips}
\alias{trim_tips}
\title{Trim tips.}
\usage{
trim_tips(path_to_ys = pkgconfig::get_config("baitfindR::path_to_ys"),
  tree_folder, tree_file_ending, relative_cutoff, absolute_cutoff,
  overwrite = FALSE, get_hash = TRUE, ...)
}
\arguments{
\item{path_to_ys}{Character vector of length one; the path to the folder containing Y&S python scripts, e.g., \code{"/Users/me/apps/phylogenomic_dataset_construction/"}}

\item{tree_folder}{Character vector of length one; the path to the folder containing the trees to trim.}

\item{tree_file_ending}{Character vector of length one; only tree files with this file ending will be used.}

\item{relative_cutoff}{Numeric vector of length one; tips on a branch 10 times longer than their sister AND longer than this value will be trimmed.}

\item{absolute_cutoff}{Numeric vector of length one; tips on branches longer than this value will be trimmed.}

\item{overwrite}{Logical; should previous output of this command be erased so new output can be written? Once erased it cannot be restored, so use with caution!}

\item{get_hash}{Logical; should the 32-byte MD5 hash be computed for all output trimmed tree files concatenated together? Used for by \code{\link[drake]{drake_plan}} for tracking during workflows. If \code{TRUE}, this function will return the hash.}

\item{...}{Other arguments. Not used by this function, but meant to be used by \code{\link[drake]{drake_plan}} for tracking during workflows.}
}
\value{
For each input tree with a file ending matching \code{tree_file_ending} in \code{tree_folder}, a trimmed tree with a file ending in \code{.tt} will be written to \code{tree_folder}. If \code{get_hash} is \code{TRUE}, the 32-byte MD5 hash be computed for all trimmed tree files concatenated together will be returned.
}
\description{
Given a folder containing phylogenetic trees, exclude (i.e., "trim"),
tips on unusually long branches. Tips on a branch 10 times longer
than their sister AND longer than \code{relative_cutoff}, OR tips
that are longer than \code{absolute_cutoff} will be trimmed. This
function will overwrite any output files with the same name in
\code{tree_folder}.
}
\details{
Wrapper for Yang and Smith (2014) \code{trim_tips.py}
}
\examples{
\dontrun{trim_tips(tree_folder = "some/folder/containing/tree/files", tree_file_ending = ".tre", relative_cutoff = 0.2, absolute_cutoff = 0.4)}
}
\references{
Yang, Y. and S.A. Smith. 2014. Orthology inference in non-model organisms using transcriptomes and low-coverage genomes: improving accuracy and matrix occupancy for phylogenomics. Molecular Biology and Evolution 31:3081-3092. \url{https://bitbucket.org/yangya/phylogenomic_dataset_construction/overview}
}
\author{
Joel H Nitta, \email{joelnitta@gmail.com}
}
